Updatinf scripts in package.json
