+++ src/controllers/accessController.ts
@@ -1,16 +1,19 @@
-import { checkRole } from '../utils/checkRole';
 import {Request, Response} from 'express';
-import checkAbilities from '../utils/checkAbilities';
+import test from '../authModules/test';
 
 function accessController(req: Request, res: Response){
     console.log("Request:", req.body);
-    const { eventType, component, credentials } = req.body;
+    const { eventType, component, credentials,  authService} = req.body;
+    
+    switch(authService){
+        case "test":
+            test(res, credentials, eventType, component);
+            break;
+        case "keycloack":
 
-    const role = checkRole(credentials.type, credentials.id);
-    console.log("role:", role);
-    const response = checkAbilities(role, eventType, component);
-    console.log("Permission response:", response);
-    res.json(response);
+    }
+
+
 }
 
 export default accessController;
